Some project reorganizing
