+++ DiscordBot/Features/Voting/Services/MongoVotesStore.cs
