+++ DiscordBot/Features/Voting/Services/VotingCooldownManager.cs
